# backend/cmd/server/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "zzrpg Backend API",
    "description": "Modern, scalable browser-based online RPG backend API documentation.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": { "type": "string", "example": "singo" },
                  "email": { "type": "string", "example": "singo@test.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": { "type": "integer", "example": 1 },
                        "username": { "type": "string", "example": "singo" },
                        "email": { "type": "string", "example": "singo@test.com" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed / Invalid request" },
          "409": { "description": "Username or email already exists" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Log in and obtain a JWT Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "singo" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string", "example": "eyJhbGciOiJIUzI1Ni..." },
                        "expires_in": { "type": "integer", "example": 86400 }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "Get authenticated user profile details",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": { "type": "integer", "example": 1 },
                        "username": { "type": "string", "example": "singo" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized / Missing or invalid token" }
        }
      }
    },
    "/api/v1/characters": {
      "get": {
        "summary": "List all characters owned by the user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of characters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer", "example": 1 },
                          "user_id": { "type": "integer", "example": 1 },
                          "name": { "type": "string", "example": "SavasciKral" },
                          "class_name": { "type": "string", "example": "WARRIOR" },
                          "level": { "type": "integer", "example": 1 },
                          "experience": { "type": "integer", "example": 0 },
                          "gold": { "type": "integer", "example": 0 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new character",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "class_name"],
                "properties": {
                  "name": { "type": "string", "example": "SavasciKral" },
                  "class_name": { "type": "string", "example": "WARRIOR" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Character created successfully with initial base and derived stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "SavasciKral" },
                        "class_name": { "type": "string", "example": "WARRIOR" },
                        "level": { "type": "integer", "example": 1 },
                        "stats": {
                          "type": "object",
                          "properties": {
                            "base_stats": {
                              "type": "object",
                              "properties": {
                                "STR": { "type": "number", "example": 15 },
                                "CON": { "type": "number", "example": 15 },
                                "INT": { "type": "number", "example": 5 },
                                "DEX": { "type": "number", "example": 10 }
                              }
                            },
                            "derived_stats": {
                              "type": "object",
                              "properties": {
                                "HP": { "type": "number", "example": 225 },
                                "MP": { "type": "number", "example": 50 },
                                "ATTACK": { "type": "number", "example": 30 },
                                "DEFENSE": { "type": "number", "example": 15 },
                                "CRIT_RATE": { "type": "number", "example": 5 }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed / Invalid class name" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Character limit reached (Max 4)" },
          "409": { "description": "Character name already taken" }
        }
      }
    },
    "/api/v1/characters/{id}": {
      "get": {
        "summary": "Get character details and calculated stats",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Character ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Character details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "SavasciKral" },
                        "class_name": { "type": "string", "example": "WARRIOR" },
                        "stats": {
                          "type": "object",
                          "properties": {
                            "base_stats": { "type": "object" },
                            "derived_stats": { "type": "object" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Character not found" }
        }
      }
    },
    "/api/v1/characters/{id}/stats": {
      "get": {
        "summary": "Get character base and derived stats",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Character ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Calculated stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "character_id": { "type": "integer", "example": 1 },
                        "base_stats": { "type": "object" },
                        "derived_stats": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Character not found" }
        }
      }
    },
    "/api/v1/admin/items": {
      "get": {
        "summary": "List all item definitions",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of item templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "dragon_sword_0" },
                          "name": { "type": "string", "example": "Dragon Sword +0" },
                          "slot_type": { "type": "string", "example": "WEAPON" },
                          "min_level": { "type": "integer", "example": 10 }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new item definition",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "name", "slot_type"],
                "properties": {
                  "id": { "type": "string", "example": "dragon_sword_0" },
                  "name": { "type": "string", "example": "Dragon Sword +0" },
                  "description": { "type": "string", "example": "A sword forged in dragon fire." },
                  "slot_type": { "type": "string", "example": "WEAPON" },
                  "min_level": { "type": "integer", "example": 10 },
                  "stats_modifiers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "stat": { "type": "string", "example": "ATTACK" },
                        "operation": { "type": "string", "example": "ADD" },
                        "value": { "type": "number", "example": 120.0 },
                        "priority": { "type": "integer", "example": 20 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Item template created successfully" },
          "400": { "description": "Validation failed" },
          "409": { "description": "Item ID already exists" }
        }
      }
    },
    "/api/v1/admin/items/{id}": {
      "get": {
        "summary": "Get item definition details",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Item template details" },
          "404": { "description": "Item not found" }
        }
      },
      "put": {
        "summary": "Update an existing item definition",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "slot_type"],
                "properties": {
                  "name": { "type": "string", "example": "Dragon Sword +1" },
                  "description": { "type": "string", "example": "Upgraded sword" },
                  "slot_type": { "type": "string", "example": "WEAPON" },
                  "min_level": { "type": "integer", "example": 10 },
                  "stats_modifiers": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Item template updated successfully" },
          "404": { "description": "Item not found" }
        }
      },
      "delete": {
        "summary": "Delete an item definition",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Item template deleted successfully" },
          "404": { "description": "Item not found" }
        }
      }
    },
    "/api/v1/characters/{id}/inventory": {
      "get": {
        "summary": "Get character inventory (bag and equipped items)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer", "example": 12 },
                          "character_id": { "type": "integer", "example": 1 },
                          "slot_index": { "type": "integer", "example": 1000 },
                          "item_definition_id": { "type": "string", "example": "dragon_sword_0" },
                          "quantity": { "type": "integer", "example": 1 },
                          "durability": { "type": "integer", "example": 100 }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/move": {
      "post": {
        "summary": "Move or equip/unequip an item",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["character_id", "from_slot", "to_slot"],
                "properties": {
                  "character_id": { "type": "integer", "example": 1 },
                  "from_slot": { "type": "integer", "example": 5 },
                  "to_slot": { "type": "integer", "example": 1000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item moved or equipped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "refresh_stats": { "type": "boolean", "example": true }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Slot index mismatch or out of bounds" },
          "403": { "description": "Level or Class restriction prevents equipping" },
          "404": { "description": "Item not found in from_slot" }
        }
      }
    },
    "/api/v1/admin/inventory/add": {
      "post": {
        "summary": "Admin: Spawn a new item in character bag",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["character_id", "item_definition_id"],
                "properties": {
                  "character_id": { "type": "integer", "example": 1 },
                  "item_definition_id": { "type": "string", "example": "dragon_sword_0" },
                  "quantity": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Item spawned successfully" }
        }
      }
    },
...
    "/api/v1/quests": {
      "get": {
        "summary": "Get all available quest definitions",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of quest definitions" }
        }
      }
    },
    "/api/v1/characters/{id}/quests/accept": {
      "post": {
        "summary": "Accept a quest for character",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quest_id"],
                "properties": {
                  "quest_id": { "type": "string", "example": "wolf_slayer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Quest accepted successfully" }
        }
      }
    },
    "/api/v1/characters/{id}/quests": {
      "get": {
        "summary": "Get active and completed quests log for character",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Character quest log details" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}